scripts/audio-generator/pyproject.toml
[project]
name = "audio-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "elevenlabs>=2.5.0",
    "python-dotenv>=1.1.1",
]
